Migrate `SearchAWSSvc` from `search` to `awscontroller` package
